extract AbstractImageMojo
